Update migrate-slack-to-teams.md
The free solution described in this article is no longer maintained.  The updated link is a fork/rewrite of the previous solution but includes additional features including the ability to migrate messages into your channels.  This solution is based on new Microsoft guidance and follows this article: https://devblogs.microsoft.com/microsoft365dev/migrate-messages-from-other-chat-platforms-to-microsoft-teams/